Update hipache atomicly
The commands issued to redis to update the hipache configuration
were all send one after another. Using a pipeline removes a
race between a request to hipache and the update, where the
client may get an error even if working upstreams are available,
but not yet added to redis again.