Fix for issue #294, Make ordering of containers in the UI consistent
Updated app/controller.js to have a default ordering of id for containers
and name for engines.  Introduced a scope variable to store whether the
sorting should be reversed.

I have then updated the engines and containers templates so that when a user
clicks on a heading, the table is sorted based upon the selected field.

When doing this, adding the <a> tag made the headings blue, so I've made a
modification to the CSS so that "th -> a" tags are black to remain consistent.